# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Stockfish Chess Analysis Service API",
...
    "version": "1.0.0"
  },
  "paths": {
...
    "/sf/v1/analyze-move": {
      "post": {
        "tags": [
          "Stockfish Analysis"
        ],
        "summary": "Analyze Move",
        "description": "一次完成最佳走法、第二候选与实战走法的同深度对拍。",
        "operationId": "analyze_move_sf_v1_analyze_move_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyzeMoveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalyzeMoveResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/sf/v1/evaluate": {
      "post": {
        "tags": [
          "Stockfish Analysis"
        ],
        "summary": "Evaluate",
        "description": "返回任意局面的综合评价与候选走法列表。",
        "operationId": "evaluate_sf_v1_evaluate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvaluateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/sf/v1/analyze-game": {
      "post": {
        "tags": [
          "Stockfish Analysis"
        ],
        "summary": "Review Game",
        "description": "对整局人类对局进行全盘高并发多核并行复盘，返回每步同深度对比、评语与全局统计。",
        "operationId": "review_game_sf_v1_analyze_game_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewGameRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewGameResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/sf/v1/review": {
      "post": {
        "tags": [
          "Stockfish Analysis"
        ],
        "summary": "Review Game",
        "description": "对整局人类对局进行全盘高并发多核并行复盘，返回每步同深度对比、评语与全局统计。",
        "operationId": "review_game_sf_v1_review_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewGameRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewGameResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AnalyzeMoveRequest": {
        "properties": {
          "requestId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requestid",
            "description": "请求唯一标识，便于跟踪或取消。"
          },
          "position": {
            "$ref": "#/components/schemas/PositionInput",
            "description": "待分析着法落子前的局面历史。"
          },
          "playedMove": {
            "type": "string",
            "title": "Playedmove",
            "description": "实战走出的单步 UCI 着法（如 f1c4）。"
          },
          "profile": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Profile",
            "description": "预设档位（lightning / fast / standard / deep / ultra）。",
            "default": "deep"
          },
          "limits": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MoveAnalysisLimits"
              },
              {
                "type": "null"
              }
            ],
            "description": "搜索预算限制；若省略则继承 profile 或标准深度 128。"
          },
          "multiPv": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 10.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Multipv",
            "description": "首阶段多候选线数（若省略则继承 profile：lightning 默认为 1，其余默认为 2）。"
          },
          "maxPvPlies": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 64.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Maxpvplies",
            "description": "返回 PV 的最大步数截断（默认 12）。",
            "default": 12
          }
        },
        "type": "object",
        "required": [
          "position",
          "playedMove"
        ],
        "title": "AnalyzeMoveRequest"
      },
      "AnalyzeMoveResponse": {
        "properties": {
          "requestId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requestid",
            "description": "请求标识"
          },
          "best": {
            "$ref": "#/components/schemas/EvalItem",
            "description": "引擎分析得出的最佳走法"
          },
          "played": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EvalItem"
              },
              {
                "type": "null"
              }
            ],
            "description": "实战走出的走法评估"
          },
          "second": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EvalItem"
              },
              {
                "type": "null"
              }
            ],
            "description": "第二候选走法，用于评估最佳走法是否明显突出"
          },
          "previousBest": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EvalItem"
              },
              {
                "type": "null"
              }
            ],
            "description": "上一完成深度（depth-1）的最佳走法，用于评估评分稳定性"
          },
          "comparison": {
            "$ref": "#/components/schemas/ComparisonResult",
            "description": "同深度对比结果"
          },
          "opening": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OpeningInfo"
              },
              {
                "type": "null"
              }
            ],
            "description": "开局理论名称与体系识别信息"
          },
          "engine": {
            "additionalProperties": true,
            "type": "object",
            "title": "Engine",
            "description": "引擎版本与运行配置"
          },
          "stats": {
            "$ref": "#/components/schemas/EngineStats",
            "description": "耗时与搜索统计"
          }
        },
        "type": "object",
        "required": [
          "best",
          "comparison",
          "engine",
          "stats"
        ],
        "title": "AnalyzeMoveResponse"
      },
...
      "EvaluateRequest": {
        "properties": {
          "requestId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requestid",
            "description": "请求唯一标识。"
          },
          "position": {
            "$ref": "#/components/schemas/PositionInput",
            "description": "待评估局面。"
          },
          "profile": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Profile",
            "description": "预设档位（lightning / fast / standard / deep / ultra）。",
            "default": "standard"
          },
          "limits": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MoveAnalysisLimits"
              },
              {
                "type": "null"
              }
            ],
            "description": "搜索限制。"
          },
          "multiPv": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 10.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Multipv",
            "description": "候选线数（若省略则继承 profile）。"
          },
          "maxPvPlies": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 64.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Maxpvplies",
            "description": "PV 步数。",
            "default": 12
          }
        },
        "type": "object",
        "required": [
          "position"
        ],
        "title": "EvaluateRequest"
      },
      "EvaluateResponse": {
        "properties": {
          "requestId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requestid",
            "description": "请求标识"
          },
          "completedDepth": {
            "type": "integer",
            "title": "Completeddepth",
            "description": "实际完成深度"
          },
          "best": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EvalItem"
              },
              {
                "type": "null"
              }
            ],
            "description": "最佳走法"
          },
          "candidates": {
            "items": {
              "$ref": "#/components/schemas/EvalItem"
            },
            "type": "array",
            "title": "Candidates",
            "description": "各候选走法评分排序列表"
          },
          "opening": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OpeningInfo"
              },
              {
                "type": "null"
              }
            ],
            "description": "开局理论名称与体系识别信息"
          },
          "engine": {
            "additionalProperties": true,
            "type": "object",
            "title": "Engine",
            "description": "引擎版本与运行配置"
          },
          "stats": {
            "$ref": "#/components/schemas/EngineStats",
            "description": "耗时与搜索统计"
          }
        },
        "type": "object",
        "required": [
          "completedDepth",
          "engine",
          "stats"
        ],
        "title": "EvaluateResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PositionInput": {
        "properties": {
          "initialFen": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Initialfen",
            "description": "起始 FEN，缺省为国际象棋标准起始局面。"
          },
          "moves": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Moves",
            "description": "从 initialFen 到当前分析局面之前的全部 UCI 历史走法列表（如 ['e2e4', 'e7e5']）。"
          }
        },
        "type": "object",
        "title": "PositionInput"
      },
      "ReviewGameRequest": {
        "properties": {
          "requestId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requestid",
            "description": "请求唯一标识，便于跟踪或取消。"
          },
          "initialFen": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Initialfen",
            "description": "起始 FEN，缺省为国际象棋标准起始局面。"
          },
          "moves": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Moves",
            "description": "从起始局面开始的全部 UCI 历史走法序列（如 ['e2e4', 'e7e5', ...]）。"
          },
          "pgn": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pgn",
            "description": "可选 PGN 文本格式棋谱。若传入且 moves 为空，则自动解析其中的 moves 与 FEN。"
          },
          "profile": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Profile",
            "description": "预设档位（默认 lightning：目标深度 22，极速并行复盘）。",
            "default": "lightning"
          },
          "concurrency": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 20.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Concurrency",
            "description": "并发分析工作引擎进程数（若省略则自适应，20核主机默认分配 16 路单线程并行加速）。"
          }
        },
        "type": "object",
        "title": "ReviewGameRequest"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "通过 Authorization: Bearer <token> 鉴权"
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Access-Token",
        "description": "通过 X-Access-Token: <token> 鉴权"
      }
    }
  },
  "servers": [
    {
      "url": "https://chess.jeefy.top",
      "description": "UniChess 线上生产服务器"
    },
    {
      "url": "http://127.0.0.1:8000",
      "description": "本地开发 / 直连服务"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    },
    {
      "ApiKeyAuth": []
    }
  ]
}